requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
numpy==1.21.2
Pillow==8.3.2
requests==2.26.0
urllib3==1.26.7
